Update part-2-using-forms.md
The current regex doesn't match more than one word limiting product names to strings without spaces. This fix updates the suggested regex to match only letters but for multiple words, and adds pattern matching for dashes as support product names that match early entries in the DB. It also does not allow a space as the first character which prevents blank entries.

The fix also updates the error message to be consistent with the above regex pattern.