Add instructions for existing deployments
The problem is that `Dockerfile` in the repository is only a wrapper over `appsmith/appsmith-ce` docker image. If the local image is outdated, docker will not by default pull a fresh one, thus current instructions will actually not lead to the AppSmith instance on Heroku being updated. If you pull image manually before executing these steps, it works fine.