Add a drafts folder
A drafts folder will hold the first drafts shared by engineers and the initial draft created by the documentation team. It will also be used to complete the reviews (EM/PM/Technical/Peers). The final version of the doc will then be ported to GitBook by the doc team member.