feat: Use absolute paths in links to internal pages (#579)
The relative paths with long `../` chains are hard to grok and easy to
get wrong. They also need fixing when a page is moved from one folder to
another. They also need fixing when a page is converted to a section,
like moving `document.md` to `document/README.md`. Docusaurus lets us
use absolute paths from the `website/docs` folder, which is very
convenient and the links are easy to read. They also don't need fixing
when the page is moved to another section.

I'm not looking to change _all_ links in the docs to use this, but am
using this PR to showcase this so that we can use absolute paths from
here on please?