chore: add script to update the index used by the omnibar (#564)
Related to https://github.com/appsmithorg/appsmith/issues/17501

Adds a script to update the index used by the omnibar.

The script fetches all the markdown under docs/ folder and parses,
uploads the same. It updates the href of links in the markdown to full
paths. The images for the omnibar are fetched from
`https://raw.githubusercontent.com/appsmithorg/appsmith-docs/main/website/static`.
The title is the h1 text in the markdown.

Also updates the workflow to run the script.

Co-authored-by: Shrikant Sharat Kandula <shrikant@appsmith.com>