fix: Notes for lack of HS256 support (#893)
Adding a note for the lack of support of the HS256 algorithm in OIDC
token signing.

I've added the note in each IDP page subtly because this is an edge case
setup. I'm worried that users might get unnecessarily distracted if we
show the note promptly or on the OIDC page.