fix: Added how-to guide for uploading a file to Dropbox (#1328)
Adds a How-To page that describes how to configure a datasource+query to
upload a file to Dropbox using OAuth 2.0 for authentication.

~~This page lives under **Reference > Datasources > Authenticated API >
How to…** because it’s a set of instructions that are used specifically
with the Auth API datasource / OAuth.~~

~~This could definitely also go under the **How-To Guides** section of
the docs though, for obvious reasons. (It just seems like we’re moving
away from this organization style).~~

This page lives in the **How-To Guides** section.

Review changes:
- Changed title: How to upload files to dropbox
- Move to how-to guides section
- fixed ugly bolding in sequence instructions for finding config values
- Reiterated the base URL when talking about configuring queries
- more specific headings "configure the filepicker" "Configure the
upload query"
- removed button from widgets setup, uses filepicker event to submit
instead
- uses UI property pane to set up event/callbacks, instead of a code
snippet

Resolves #1305 

## Checklist
I have:
- [x] run the content through Grammarly
- [x] linked to sample apps when relevant
- [x] added the meta description for each page in the PR
- [x] minimized the callouts and added only when necessary
- [x] added the `queryString` parameter to the Tabs (if used)
- [x] masked PII in images. For example, login credentials, account
details, and more
- [x] added images only when necessary
- [x] deleted the images that are no longer used for the updated pages
in the PR
- [x] followed the image file naming convention while renaming or adding
new images. (Use lowercase letters, dashes between words, and be as
descriptive as possible)
- [x] used the `<figure/>` tag instead of a markdown representation for
images
- [x] added the `<figcaption/>` tag to add a caption to the image
- [x] added the `alt` attribute in the `<img/>` tag
- [x] verified and updated the cross-references or created redirect
rules for the changed or removed content
- [x] reviewed and applied the style changes for UI formatting. For
example, Bold the UI elements(Buttons on screen) used in the doc.