fix: update Kubernetes upgrade instructions to explicitly set image tag (#2903)
## Description 

I'm updating the Kubernetes upgrade steps to explicitly set the version
that should be used.

The old method assumed that the administrator was installing Appsmith in
a non-HA installation and they actually may or may not actually trigger
an upgrade in the default situation: it depends upon whether the new pod
gets scheduled on a new node or not. It could even trigger a downgrade
if the pod got scheduled on a node that happened to have an older copy
of the Appsmith image.

I'm going to open a separate ticket to revamp the installation
instructions for Kubernetes as using `latest` can lead to some
unexpected situations depending upon how the cluster is setup.

## Pull request type

Check the appropriate box:

- [ ] Review Fixes
- [x] Documentation Overhaul
- [ ] Feature/Story
    - Link one or more Engineering Tickets
        * 
- [ ] A-Force
- [ ] Error in documentation
- [ ] Maintenance

## Documentation tickets

 Link to one or more documentation tickets:
 - 

## Checklist

From the below options, select the ones that are applicable:

- [ ] Checked for Grammarly suggestions.
- [ ] Adhered to the writing checklist.
- [ ] Adhered to the media checklist.
- [ ] Verified and updated cross-references or added redirect rules.
- [ ] Tested the redirect rules on deploy preview.
- [ ] Validated the modifications made to the content on the deploy
preview.
- [ ] Validated the CSS modifications on different screen sizes.